pyproject.toml
[tool.poetry]
name = "hs-exchange-reader"
version = "0.1.0"
description = "Async exchange symbol sync and reader service"
authors = ["Bolong He <you@example.com>"]
package-mode = false
packages = [
    { include = "databases" },
    { include = "exchanges" },
    { include = "jobs" },
    { include = "utils" },
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
SQLAlchemy = { version = "2.0.44", extras = ["asyncio"] }
apscheduler = "3.11.0"
requests = { extras = ["socks"], version = "*" }
python-dotenv = "1.1.1"
loguru = "0.7.3"
pymysql = "1.1.2"
cryptography = "46.0.3"
aiomysql = "0.3.2"
aiohttp = "3.13.2"
pandas = "2.3.3"
structlog = "^25.5.0"
python-json-logger = "^4.0.0"
yfinance = "^0.2.66"
pycryptodome = "^3.23.0"
[tool.poetry.group.dev.dependencies]
ruff = ">=0.6.0"
black = ">=24.4.0"
mypy = ">=1.10.0"
types-requests = "*"
pytest = ">=8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ===============================
# Ruff 配置
# ===============================
[tool.ruff]
target-version = "py313"
line-length = 120
src = ["databases", "exchanges", "jobs", "utils"]

select = [
    "E", "W", "F", "I", "UP", "B", "A", "C4", "TID", "N", "ASYNC", "RUF"
]

ignore = [
    "E501",
    "B008",
    "N818",
]

exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".pytest_cache",
    "__pycache__",
    "venv",
    "build",
    "dist",
]

[tool.ruff.isort]
known-first-party = ["databases", "exchanges", "jobs", "utils"]
combine-as-imports = true
force-sort-within-sections = true
lines-between-types = 1
order-by-type = true

[tool.ruff.pyupgrade]
keep-runtime-typing = true

[tool.ruff.per-file-ignores]
"tests/*" = ["D", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
docstring-code-format = true
docstring-code-line-length = 88

# ===============================
# MyPy 配置
# ===============================
[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
disallow_untyped_defs = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
pretty = true
show_error_codes = true
check_untyped_defs = true
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
